@@ -11,7 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 """Holds the ModelBuilder class and the ModelServer enum."""
-from __future__ import absolute_import
+from __future__ import absolute_import, annotations
 
 import importlib.util
 import json
@@ -24,6 +24,7 @@ import re
 
 from pathlib import Path
 
+from botocore.exceptions import ClientError
 from sagemaker_core.main.resources import TrainingJob
 
 from sagemaker.transformer import Transformer
@@ -37,6 +38,7 @@ from sagemaker.jumpstart.utils import get_jumpstart_content_bucket
 from sagemaker.s3 import S3Downloader
 from sagemaker import Session
 from sagemaker.model import Model
+from sagemaker.jumpstart.model import JumpStartModel
 from sagemaker.base_predictor import PredictorBase
 from sagemaker.serializers import NumpySerializer, TorchTensorSerializer
 from sagemaker.deserializers import JSONDeserializer, TorchTensorDeserializer
@@ -75,6 +77,7 @@ from sagemaker.serve.model_format.mlflow.utils import (
 )
 from sagemaker.serve.save_retrive.version_1_0_0.metadata.metadata import Metadata
 from sagemaker.serve.spec.inference_spec import InferenceSpec
+from sagemaker.serve.spec.inference_base import CustomOrchestrator, AsyncCustomOrchestrator
 from sagemaker.serve.utils import task
 from sagemaker.serve.utils.exceptions import TaskNotFoundException
 from sagemaker.serve.utils.lineage_utils import _maintain_lineage_tracking_for_mlflow_model
@@ -102,6 +105,7 @@ from sagemaker.serve.detector.image_detector import (
     _get_model_base,
 )
 from sagemaker.serve.model_server.torchserve.prepare import prepare_for_torchserve
+from sagemaker.serve.model_server.smd.prepare import prepare_for_smd
 from sagemaker.serve.model_server.triton.triton_builder import Triton
 from sagemaker.serve.utils.telemetry_logger import _capture_telemetry
 from sagemaker.serve.utils.types import ModelServer, ModelHub
@@ -131,6 +135,7 @@ supported_model_servers = {
     ModelServer.MMS,
     ModelServer.TGI,
     ModelServer.TEI,
+    ModelServer.SMD,
 }
 
 
@@ -220,6 +225,18 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
             available for providing s3 path to fine-tuned model artifacts. ``FINE_TUNING_JOB_NAME``
             is available for providing fine-tuned job name. Both ``FINE_TUNING_MODEL_PATH`` and
             ``FINE_TUNING_JOB_NAME`` are mutually exclusive.
+        inference_component_name (Optional[str]): The name for an inference component
+            created from this ModelBuilder instance. This or ``resource_requirements`` must be set
+            to denote that this instance refers to an inference component.
+        modelbuilder_list: Optional[List[ModelBuilder]] = List of ModelBuilder objects which
+            can be built in bulk and subsequently deployed in bulk. Currently only supports
+            deployments for inference components.
+        resource_requirements: Optional[ResourceRequirements] = Defines the compute resources
+            allocated to run the model assigned to the inference component. This or
+            ``inference_component_name`` must be set to denote that this instance refers
+            to an inference component. If ``inference_component_name`` is set but this is not and a
+            JumpStart model ID is specified, pre-benchmarked deployment configs will attempt to be
+            retrieved for the model.
     """
 
     model_path: Optional[str] = field(
@@ -233,7 +250,7 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
         default=None, metadata={"help": "Define sagemaker session for execution"}
     )
     name: Optional[str] = field(
-        default="model-name-" + uuid.uuid1().hex,
+        default_factory=lambda: "model-name-" + uuid.uuid1().hex,
         metadata={"help": "Define the model name"},
     )
     mode: Optional[Mode] = field(
@@ -318,6 +335,23 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
             "`MLFLOW_MODEL_PATH`, `FINE_TUNING_MODEL_PATH`, `FINE_TUNING_JOB_NAME`, and "
             "`CUSTOM_MODEL_PATH`. HF_TASK should be set for new models without task metadata "
             "in the Hub, Adding unsupported task types will throw an exception."
+        },
+    )
+    inference_component_name: Optional[str] = field(
+        default=None,
+        metadata={
+            "help": "Defines the name for an Inference Component created from this ModelBuilder."
+        },
+    )
+    modelbuilder_list: Optional[List[ModelBuilder]] = field(
+        default=None,
+        metadata={"help": "Defines a list of ModelBuilder objects."},
+    )
+    resource_requirements: Optional[ResourceRequirements] = field(
+        default=None,
+        metadata={
+            "help": "Defines the compute resources allocated to run the model assigned"
+            " to the inference component."
         },
     )
 
@@ -465,7 +499,7 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
         elif self.schema_builder:
             serializer = self.schema_builder.input_serializer
         else:
-            raise Exception("Cannot serialize")
+            raise Exception("Cannot serialize. Try providing a SchemaBuilder if not present.")
 
         deserializer = None
         if self.accept_type == "application/json":
@@ -477,7 +511,7 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
         elif self.schema_builder:
             deserializer = self.schema_builder.output_deserializer
         else:
-            raise Exception("Cannot deserialize")
+            raise Exception("Cannot deserialize. Try providing a SchemaBuilder if not present.")
 
         return serializer, deserializer
 
@@ -562,6 +596,83 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
         self.pysdk_model.model_package_arn = None
         return predictor
 
+    def _deploy_for_ic(
+        self,
+        *args,
+        ic_data: Dict[str, Any],
+        container_timeout_in_seconds: int = 300,
+        model_data_download_timeout: int = 3600,
+        instance_type: Optional[str] = None,
+        initial_instance_count: Optional[int] = None,
+        endpoint_name: Optional[str] = None,
+        **kwargs,
+    ) -> Predictor:
+        """Creates an Inference Component from a ModelBuilder."""
+        ic_name = ic_data.get("Name", None)
+        model = ic_data.get("Model", None)
+        resource_requirements = ic_data.get("ResourceRequirements", {})
+
+        # Ensure resource requirements are set for non-JumpStart models
+        if not resource_requirements:
+            raise ValueError(
+                f"Cannot create/update inference component {ic_name} without resource requirements."
+            )
+
+        # Check if the Inference Component exists
+        if ic_name and self._does_ic_exist(ic_name=ic_name):
+            logger.info("Updating Inference Component %s as it already exists.", ic_name)
+
+            # Create spec for updating the IC
+            startup_parameters = {}
+            if model_data_download_timeout is not None:
+                startup_parameters["ModelDataDownloadTimeoutInSeconds"] = (
+                    model_data_download_timeout
+                )
+            if container_timeout_in_seconds is not None:
+                startup_parameters["ContainerStartupHealthCheckTimeoutInSeconds"] = (
+                    container_timeout_in_seconds
+                )
+            compute_rr = resource_requirements.get_compute_resource_requirements()
+            inference_component_spec = {
+                "ModelName": self.name,
+                "StartupParameters": startup_parameters,
+                "ComputeResourceRequirements": compute_rr,
+            }
+            runtime_config = {"CopyCount": resource_requirements.copy_count}
+            response = self.sagemaker_session.update_inference_component(
+                inference_component_name=ic_name,
+                specification=inference_component_spec,
+                runtime_config=runtime_config,
+            )
+            return Predictor(endpoint_name=response.get("EndpointName"), component_name=ic_name)
+        else:
+            kwargs.update(
+                {
+                    "resources": resource_requirements,
+                    "endpoint_type": EndpointType.INFERENCE_COMPONENT_BASED,
+                    "inference_component_name": ic_name,
+                    "endpoint_logging": False,
+                }
+            )
+            return model.deploy(
+                *args,
+                container_startup_health_check_timeout=container_timeout_in_seconds,
+                initial_instance_count=initial_instance_count,
+                instance_type=instance_type,
+                mode=Mode.SAGEMAKER_ENDPOINT,
+                endpoint_name=endpoint_name,
+                **kwargs,
+            )
+
+    def _does_ic_exist(self, ic_name: str) -> bool:
+        """Returns true if an Inference Component exists with the given name."""
+        try:
+            self.sagemaker_session.describe_inference_component(inference_component_name=ic_name)
+            return True
+        except ClientError as e:
+            msg = e.response["Error"]["Message"]
+            return "Could not find inference component" not in msg
+
     @_capture_telemetry("torchserve.deploy")
     def _model_builder_deploy_wrapper(
         self,
@@ -615,6 +726,13 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
 
         if "endpoint_logging" not in kwargs:
             kwargs["endpoint_logging"] = True
+
+        if "inference_component_name" not in kwargs and self.inference_component_name:
+            kwargs["inference_component_name"] = self.inference_component_name
+
+        if "resources" not in kwargs and self.resource_requirements:
+            kwargs["resources"] = self.resource_requirements
+
         kwargs.pop("mode", None)
         self.pysdk_model.role = kwargs.pop("role", self.pysdk_model.role)
         predictor = self._original_deploy(
@@ -666,6 +784,24 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
                 dependencies=self.dependencies,
                 session=self.sagemaker_session,
                 image_uri=self.image_uri,
+                inference_spec=self.inference_spec,
+            )
+
+        self._prepare_for_mode()
+        self.model = self._create_model()
+        return self.model
+
+    def _build_for_smd(self) -> Type[Model]:
+        """Build the model for SageMaker Distribution"""
+        self._save_model_inference_spec()
+
+        if self.mode != Mode.IN_PROCESS:
+            self._auto_detect_container()
+
+            self.secret_key = prepare_for_smd(
+                model_path=self.model_path,
+                shared_libs=self.shared_libs,
+                dependencies=self.dependencies,
                 inference_spec=self.inference_spec,
             )
 
@@ -854,13 +990,225 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
         """Dummy method to collect telemetry for estimator handshake"""
         return
 
+    def build(
+        self,
+        mode: Type[Mode] = None,
+        role_arn: str = None,
+        sagemaker_session: Optional[Session] = None,
+    ) -> Union[ModelBuilder, Type[Model]]:
+        """Creates deployable ``Model`` instances with all provided ``ModelBuilder`` objects.
+
+        Args:
+            mode (Type[Mode], optional): The mode. Defaults to ``None``.
+            role_arn (str, optional): The IAM role arn. Defaults to ``None``.
+            sagemaker_session (Optional[Session]): Session object which manages interactions
+                with Amazon SageMaker APIs and any other AWS services needed. If not specified, the
+                function creates one using the default AWS configuration chain.
+
+        Returns:
+            Union[ModelBuilder, Type[Model]]: A deployable ``ModelBuilder`` object if multiple
+            ``ModelBuilders`` were built, or a deployable ``Model`` object.
+        """
+        if role_arn:
+            self.role_arn = role_arn
+        self.sagemaker_session = sagemaker_session or self.sagemaker_session or Session()
+
+        deployables = {}
+
+        if not self.modelbuilder_list and not isinstance(
+            self.inference_spec, (CustomOrchestrator, AsyncCustomOrchestrator)
+        ):
+            self.serve_settings = self._get_serve_setting()
+            return self._build_single_modelbuilder(
+                mode=mode,
+                role_arn=self.role_arn,
+                sagemaker_session=sagemaker_session,
+            )
+
+        # Multi-ModelBuilder case: deploy
+        built_ic_models = []
+        if self.modelbuilder_list:
+            logger.info("Detected ModelBuilders in modelbuilder_list.")
+            for mb in self.modelbuilder_list:
+                if mb.mode == Mode.IN_PROCESS or mb.mode == Mode.LOCAL_CONTAINER:
+                    raise ValueError(
+                        "Bulk ModelBuilder building is only supported for SageMaker Endpoint Mode."
+                    )
+
+                if (not mb.resource_requirements and not mb.inference_component_name) and (
+                    not mb.inference_spec
+                    or not isinstance(
+                        mb.inference_spec, (CustomOrchestrator, AsyncCustomOrchestrator)
+                    )
+                ):
+                    raise ValueError(
+                        "Bulk ModelBuilder building is only supported for Inference Components "
+                        + "and custom orchestrators."
+                    )
+
+            for mb in self.modelbuilder_list:
+                # Custom orchestrator definition found in inference_spec
+                mb.serve_settings = mb._get_serve_setting()
+                # Build for Inference Component
+                logger.info("Building ModelBuilder %s.", mb.name)
+                # Get JS deployment configs if ResourceRequirements not set
+
+                mb = mb._get_ic_resource_requirements(mb=mb)
+
+                built_model = mb._build_single_modelbuilder(
+                    role_arn=self.role_arn, sagemaker_session=self.sagemaker_session
+                )
+                built_ic_models.append(
+                    {
+                        "Name": mb.inference_component_name,
+                        "ResourceRequirements": mb.resource_requirements,
+                        "Model": built_model,
+                    }
+                )
+                logger.info(
+                    "=====================Build for %s complete.===================",
+                    mb.model,
+                )
+            deployables["InferenceComponents"] = built_ic_models
+
+        if isinstance(self.inference_spec, (CustomOrchestrator, AsyncCustomOrchestrator)):
+            logger.info("Building custom orchestrator.")
+            if self.mode == Mode.IN_PROCESS or self.mode == Mode.LOCAL_CONTAINER:
+                raise ValueError(
+                    "Custom orchestrator deployment is only supported for"
+                    "SageMaker Endpoint Mode."
+                )
+            self.serve_settings = self._get_serve_setting()
+            cpu_or_gpu_instance = self._get_processing_unit()
+            self.image_uri = self._get_smd_image_uri(processing_unit=cpu_or_gpu_instance)
+            self.model_server = ModelServer.SMD
+            built_orchestrator = self._build_single_modelbuilder(
+                mode=Mode.SAGEMAKER_ENDPOINT,
+                role_arn=role_arn,
+                sagemaker_session=sagemaker_session,
+            )
+            if not self.resource_requirements:
+                logger.info(
+                    "Custom orchestrator resource_requirements not found. "
+                    "Building as a SageMaker Endpoint instead of Inference Component."
+                )
+                deployables["CustomOrchestrator"] = {
+                    "Mode": "Endpoint",
+                    "Model": built_orchestrator,
+                }
+            else:
+                # Network isolation of ICs on an endpoint must be consistent
+                if built_ic_models:
+                    if (
+                        self.dependencies["auto"]
+                        or "requirements" in self.dependencies
+                        or "custom" in self.dependencies
+                    ):
+                        logger.warning(
+                            "Custom orchestrator network isolation must be False when dependencies "
+                            "are specified or using autocapture. To enable network isolation, "
+                            "package all dependencies in the container or model artifacts "
+                            "ahead of time."
+                        )
+                        built_orchestrator._enable_network_isolation = False
+                        for model in built_ic_models:
+                            model["Model"]._enable_network_isolation = False
+                deployables["CustomOrchestrator"] = {
+                    "Name": self.inference_component_name,
+                    "Mode": "InferenceComponent",
+                    "ResourceRequirements": self.resource_requirements,
+                    "Model": built_orchestrator,
+                }
+
+            logger.info(
+                "=====================Custom orchestrator build complete.===================",
+            )
+
+        self._deployables = deployables
+        return self
+
+    def _get_processing_unit(self):
+        """Detects if the resource requirements are intended for a CPU or GPU instance."""
+        # Assume custom orchestrator will be deployed as an endpoint to a CPU instance
+        if not self.resource_requirements or not self.resource_requirements.num_accelerators:
+            return "cpu"
+        for ic in self.modelbuilder_list or []:
+            if ic.resource_requirements.num_accelerators > 0:
+                return "gpu"
+        if self.resource_requirements.num_accelerators > 0:
+            return "gpu"
+
+        return "cpu"
+
+    def _get_ic_resource_requirements(self, mb: ModelBuilder = None) -> ModelBuilder:
+        """Attempts fetching pre-benchmarked resource requirements for the MB from JumpStart."""
+        if mb._is_jumpstart_model_id() and not mb.resource_requirements:
+            js_model = JumpStartModel(model_id=mb.model)
+            deployment_configs = js_model.list_deployment_configs()
+            if not deployment_configs:
+                raise ValueError(
+                    "No resource requirements were provided for Inference Component "
+                    f"{mb.inference_component_name} and no default deployment"
+                    " configs were found in JumpStart."
+                )
+            compute_requirements = (
+                deployment_configs[0].get("DeploymentArgs").get("ComputeResourceRequirements")
+            )
+            logger.info("Retrieved pre-benchmarked deployment configurations from JumpStart.")
+            mb.resource_requirements = ResourceRequirements(
+                requests={
+                    "memory": compute_requirements["MinMemoryRequiredInMb"],
+                    "num_accelerators": compute_requirements.get(
+                        "NumberOfAcceleratorDevicesRequired", None
+                    ),
+                    "copies": 1,
+                    "num_cpus": compute_requirements.get("NumberOfCpuCoresRequired", None),
+                },
+                limits={"memory": compute_requirements.get("MaxMemoryRequiredInMb", None)},
+            )
+
+        return mb
+
+    @_capture_telemetry("build_custom_orchestrator")
+    def _get_smd_image_uri(self, processing_unit: str = None) -> str:
+        """Gets the SMD Inference Image URI.
+
+        Returns:
+            str: SMD Inference Image URI.
+        """
+        from sagemaker import image_uris
+        import sys
+
+        self.sagemaker_session = self.sagemaker_session or Session()
+        from packaging.version import Version
+
+        formatted_py_version = f"py{sys.version_info.major}{sys.version_info.minor}"
+        if Version(f"{sys.version_info.major}{sys.version_info.minor}") < Version("3.12"):
+            raise ValueError(
+                f"Found Python version {formatted_py_version} but"
+                f"Custom orchestrator deployment requires Python version >= 3.12."
+            )
+
+        INSTANCE_TYPES = {"cpu": "ml.c5.xlarge", "gpu": "ml.g5.4xlarge"}
+
+        logger.info("Finding SMD inference image URI for a %s instance.", processing_unit)
+
+        smd_uri = image_uris.retrieve(
+            framework="sagemaker-distribution",
+            image_scope="inference",
+            instance_type=INSTANCE_TYPES[processing_unit],
+            region=self.sagemaker_session.boto_region_name,
+        )
+        logger.info("Found compatible image %s", smd_uri)
+        return smd_uri
+
     # Model Builder is a class to build the model for deployment.
     # It supports three modes of deployment
     # 1/ SageMaker Endpoint
     # 2/ Local launch with container
     # 3/ In process mode with Transformers server in beta release
     @_capture_telemetry("ModelBuilder.build")
-    def build(  # pylint: disable=R0911
+    def _build_single_modelbuilder(  # pylint: disable=R0911
         self,
         mode: Type[Mode] = None,
         role_arn: str = None,
@@ -1038,6 +1386,9 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
 
         if self.model_server == ModelServer.MMS:
             return self._build_for_transformers()
+
+        if self.model_server == ModelServer.SMD:
+            return self._build_for_smd()
 
     @_capture_telemetry("ModelBuilder.save")
     def save(
@@ -1593,6 +1944,8 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
     def deploy(
         self,
         endpoint_name: str = None,
+        container_timeout_in_second: int = 300,
+        instance_type: str = None,
         initial_instance_count: Optional[int] = 1,
         inference_config: Optional[
             Union[
@@ -1603,7 +1956,10 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
             ]
         ] = None,
         update_endpoint: Optional[bool] = False,
-    ) -> Union[Predictor, Transformer]:
+        custom_orchestrator_instance_type: str = None,
+        custom_orchestrator_initial_instance_count: int = None,
+        **kwargs,
+    ) -> Union[Predictor, Transformer, List[Predictor]]:
         """Deploys the built Model.
 
         Depending on the type of config provided, this function will call deployment accordingly.
@@ -1625,42 +1981,43 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
             Transformer for Batch Deployments
             Predictors for all others
         """
-        if not hasattr(self, "built_model"):
-            raise ValueError("Model Needs to be built before deploying")
+        if not hasattr(self, "built_model") and not hasattr(self, "_deployables"):
+            raise ValueError("Model needs to be built before deploying")
         if not update_endpoint:
             endpoint_name = unique_name_from_base(endpoint_name)
 
-        if not inference_config:  # Real-time Deployment
-            return self.built_model.deploy(
-                instance_type=self.instance_type,
-                initial_instance_count=initial_instance_count,
-                endpoint_name=endpoint_name,
-                update_endpoint=update_endpoint,
-            )
+        if not hasattr(self, "_deployables"):
+            if not inference_config:  # Real-time Deployment
+                return self.built_model.deploy(
+                    instance_type=self.instance_type,
+                    initial_instance_count=initial_instance_count,
+                    endpoint_name=endpoint_name,
+                    update_endpoint=update_endpoint,
+                )
 
-        if isinstance(inference_config, ServerlessInferenceConfig):
-            return self.built_model.deploy(
-                serverless_inference_config=inference_config,
-                endpoint_name=endpoint_name,
-                update_endpoint=update_endpoint,
-            )
+            if isinstance(inference_config, ServerlessInferenceConfig):
+                return self.built_model.deploy(
+                    serverless_inference_config=inference_config,
+                    endpoint_name=endpoint_name,
+                    update_endpoint=update_endpoint,
+                )
 
-        if isinstance(inference_config, AsyncInferenceConfig):
-            return self.built_model.deploy(
-                instance_type=self.instance_type,
-                initial_instance_count=initial_instance_count,
-                async_inference_config=inference_config,
-                endpoint_name=endpoint_name,
-                update_endpoint=update_endpoint,
-            )
+            if isinstance(inference_config, AsyncInferenceConfig):
+                return self.built_model.deploy(
+                    instance_type=self.instance_type,
+                    initial_instance_count=initial_instance_count,
+                    async_inference_config=inference_config,
+                    endpoint_name=endpoint_name,
+                    update_endpoint=update_endpoint,
+                )
 
-        if isinstance(inference_config, BatchTransformInferenceConfig):
-            transformer = self.built_model.transformer(
-                instance_type=inference_config.instance_type,
-                output_path=inference_config.output_path,
-                instance_count=inference_config.instance_count,
-            )
-            return transformer
+            if isinstance(inference_config, BatchTransformInferenceConfig):
+                transformer = self.built_model.transformer(
+                    instance_type=inference_config.instance_type,
+                    output_path=inference_config.output_path,
+                    instance_count=inference_config.instance_count,
+                )
+                return transformer
 
         if isinstance(inference_config, ResourceRequirements):
             if update_endpoint:
@@ -1678,7 +2035,61 @@ class ModelBuilder(Triton, DJL, JumpStart, TGI, Transformers, TensorflowServing,
                 update_endpoint=update_endpoint,
             )
 
-        raise ValueError("Deployment Options not supported")
+            raise ValueError("Deployment Options not supported")
+
+        # Iterate through deployables for a custom orchestrator deployment.
+        # Create all Inference Components first before deploying custom orchestrator if present.
+        predictors = []
+        for inference_component in self._deployables.get("InferenceComponents", []):
+            predictors.append(
+                self._deploy_for_ic(
+                    ic_data=inference_component,
+                    container_timeout_in_seconds=container_timeout_in_second,
+                    instance_type=instance_type,
+                    initial_instance_count=initial_instance_count,
+                    endpoint_name=endpoint_name,
+                    **kwargs,
+                )
+            )
+        if self._deployables.get("CustomOrchestrator", None):
+            custom_orchestrator = self._deployables.get("CustomOrchestrator")
+            if not custom_orchestrator_instance_type and not instance_type:
+                logger.warning(
+                    "Deploying custom orchestrator as an endpoint but no instance type was "
+                    "set. Defaulting to `ml.c5.xlarge`."
+                )
+                custom_orchestrator_instance_type = "ml.c5.xlarge"
+                custom_orchestrator_initial_instance_count = 1
+            if custom_orchestrator["Mode"] == "Endpoint":
+                logger.info(
+                    "Deploying custom orchestrator on instance type %s.",
+                    custom_orchestrator_instance_type,
+                )
+                predictors.append(
+                    custom_orchestrator["Model"].deploy(
+                        instance_type=custom_orchestrator_instance_type,
+                        initial_instance_count=custom_orchestrator_initial_instance_count,
+                        **kwargs,
+                    )
+                )
+            elif custom_orchestrator["Mode"] == "InferenceComponent":
+                logger.info(
+                    "Deploying custom orchestrator as an inference component "
+                    f"to endpoint {endpoint_name}"
+                )
+                predictors.append(
+                    self._deploy_for_ic(
+                        ic_data=custom_orchestrator,
+                        container_timeout_in_seconds=container_timeout_in_second,
+                        instance_type=custom_orchestrator_instance_type or instance_type,
+                        initial_instance_count=custom_orchestrator_initial_instance_count
+                        or initial_instance_count,
+                        endpoint_name=endpoint_name,
+                        **kwargs,
+                    )
+                )
+
+        return predictors
 
     def display_benchmark_metrics(self, **kwargs):
         """Display Markdown Benchmark Metrics for deployment configs."""
